Add support for Promise based deferred elements
