+++ test/test_render.js
@@ -1,8 +1,8 @@
 /* global describe, it */
-import { createElement, HTMLString, Fragment, DeferredElement, renderSync, renderAsync } from "../src";
+import { createElement, HTMLString, Fragment, renderSync, renderAsync } from "../src";
 import BufferedStream from "../src/stream/buffered";
 import assert, { strictEqual as assertSame } from "assert";
-import { SiteIndex, NonBlockingContainer } from "./macros";
+import { SiteIndex } from "./macros";
 
 function renderAndAssert(element, string) {
 	// Sync first
@@ -146,71 +146,3 @@ describe("HTML encoding", _ => {
 		assert(html.includes("<p>foo <i>bar</i> baz</p>"));
 	});
 });
-
-describe("synchronous rendering", () => {
-	it("should detect non-blocking child elements", done => {
-		let stream = new BufferedStream();
-		let fn = _ => renderSync(NonBlockingContainer(), stream);
-		assert.throws(fn, /deferred elements unsupported in synchronous rendering/);
-		done();
-	});
-});
-
-describe("asynchronous rendering", () => {
-	it("should support deferred child elements", () => {
-		let root = createElement("body", null, new DeferredElement(callback => {
-			setTimeout(() => {
-				let el = createElement("p", { class: "info" }, "lorem", "ipsum");
-				callback(el);
-			}, 10);
-		}));
-		return renderAsync(root, new BufferedStream()).then(stream => {
-			let html = stream.read();
-			assertSame(html, '<body><p class="info">loremipsum</p></body>');
-		});
-	});
-
-	it("should keep order if deferred child elements", () => {
-		let root = createElement("body", null,
-				new DeferredElement(callback => {
-					setTimeout(() => {
-						callback(createElement("a1"));
-					}, 10);
-				}),
-				createElement("a2", null,
-						new DeferredElement(callback => {
-							setTimeout(() => {
-								callback(createElement("a3"));
-							}, 10);
-						})),
-				createElement("a4")
-		);
-		return renderAsync(root, new BufferedStream()).then(stream => {
-			let html = stream.read();
-			assertSame(html, "<body><a1></a1><a2><a3></a3></a2><a4></a4></body>");
-		});
-	});
-
-	it("should support large numbers of deferred child elements in parallel", () => {
-		let deferred = range(10000).map((_, i) => {
-			i++;
-			return new DeferredElement(callback => {
-				setTimeout(() => {
-					let el = createElement("li", null, i);
-					callback(el);
-				}, 100); // 100ms would lead to timeout if not executed parallely
-			});
-		});
-		let root = createElement("ul", null, ...deferred);
-
-		return renderAsync(root, new BufferedStream()).then(stream => {
-			let html = stream.read();
-			let expectedLis = range(10000).map((_, i) => `<li>${i}</li>`).join("");
-			assert(html, `<ul>${expectedLis}</ul>`);
-		});
-	});
-});
-
-function range(size) {
-	return Array.apply(null, Array(size));
-}
